\name{plot.mple}
\alias{plot.mple}
\title{Show the Process for Optimizing Parameter Set
}
\description{
  Plot method for object of class \code{"mple"} shows process for optimizing the
  normalized parameters depending on a given initial guess of each parameter.
}
\usage{
\method{plot}{mple}(x, \dots)
}
\arguments{
 \item{x}{an object of class \code{"mple"} returned by \code{\link{mple.cppm}}.}
 \item{\dots}{further graphical parameters from \code{\link{par}}.}
}
\keyword{spatial}
