\name{plot.Palm}
\alias{plot.Palm}
\title{
Plot Non-Parametric and Parametric Normalized Palm Intensity
}
\description{
  Plot method for objects of class \code{"Palm"}.
}
\usage{
\method{plot}{Palm}(x, \dots, log = "xy")
}
\arguments{
  \item{x}{an object of class \code{"Palm"}, usually the result of a call to
    \code{\link{palm.cppm}}.}
  \item{\dots}{optional. At most 4 additional objects of class "Palm".}
  \item{log}{a character string indicating if logarithmic axes are to be used.}
}
\keyword{spatial}
